% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kilde.R
\name{kilde}
\alias{kilde}
\title{Kildeangivelse på diagrammer}
\usage{
kilde(x, y)
}
\arguments{
\item{x}{er et ggplot2 diagram.}

\item{y}{er en tekst string indeholdende kildeangivelsen.}
}
\description{
Funktionen gør det nemt at angive kilde på et diagram.
}
\examples{
#data eksempel (fra ggplot2 examples)
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)), y = rnorm(30))

#plot eksempel (fra ggplot2 examples)
myplot <- ggplot(df, aes(x = gp, y = y))+geom_point()

kilde(myplot, 'Kilde: Væksthus Nordjylland, September 2015')
}
\keyword{kildeangivelse}

