% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plot flere diagrammer på én side.}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...,}{ggplot objects}

\item{plotlist,}{en liste af ggplot objects}

\item{cols,}{antallet af kolonner i layoutet}

\item{layout,}{matrix som specificere layoutet. Hvis layout benyttes ignoreres 'cols'}
}
\description{
Funktionen gør det muligt at plotte flere diagrammer på samme side. Funktionen er skamløst kopieret fra http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
#data eksempel (fra ggplot2 examples)
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)), y = rnorm(30))

#plot eksempel (fra ggplot2 examples)
plot1 <- ggplot(df, aes(x = gp, y = y))+geom_point()

plot2 <- ggplot(df, aes(x = gp, y = y))+geom_boxplot()


multiplot(plot1, plot2, ncols=2)

Hvis layoutet  f.eks. er matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
så vil plot 1 være til venstre øverst, 2 vil være til højre øverst, og
3 vil sprede sig nederst.
}
\keyword{multiplot}

