% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaekst_ans.R
\name{vaekst_ans}
\alias{vaekst_ans}
\title{Udregning af vækstkategori baseret på udvikling i ansatte i de 3 seneste år}
\usage{
vaekst_ans(x, y, z, g)
}
\arguments{
\item{x}{er en numerisk variabel og indeholder virksomhedens antal ansatte  til tiden t.}

\item{y}{er en numerisk variabel og indeholder virksomhedens antal ansatte  til tiden t+1.}

\item{z}{er en numerisk variabel og indeholder virksomhedens antal ansatte  til tiden t+2.}

\item{g}{er en numerisk variabel og indeholder virksomhedens antal ansatte  til tiden t+3.}
}
\description{
Funktionen udregner vækstkategorien for udviklingen i antal ansatte
}
\details{
Kategorisering Beskedent,Moderat, Højvækst bliver kun foretaget hvis virksomheden tilhører samme kategori i alle de seneste 3 år, dvs. at f.eks. højvækst eller beskedent i år1 og 2, men moderat i år3 resultere i NA. Kategorierne eksistere altså kun i deres "reneste form"

Forudsætning for at beregning foretages er at antal ansatte er min. 5 i starten(g) af den seneste 3-årige periode
}
\examples{
#data eksempel 

x <- 8 # Antal ansatte i 2014
y <- 7 # Antal ansatte i 2013
z <- 6 # Antal ansatte i 2012
g <- 5 # Antal ansatte i 2011

vaekst_ans(x,y,z,g)
}
\keyword{vækstkategori}

