% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaekst_oms.R
\name{vaekst_oms}
\alias{vaekst_oms}
\title{Udregning af vækstkategori baseret på udvikling i omsætning i de 3 seneste år}
\usage{
vaekst_oms(x, y, z, g)
}
\arguments{
\item{x}{er en numerisk variabel og indeholder virksomhedens omsætning angivet i tusinde kroner til tiden t.}

\item{y}{er en numerisk variabel og indeholder virksomhedens omsætning angivet i tusinde kroner til tiden t+1.}

\item{z}{er en numerisk variabel og indeholder virksomhedens omsætning angivet i tusinde kroner til tiden t+2.}

\item{g}{er en numerisk variabel og indeholder virksomhedens omsætning angivet i tusinde kroner til tiden t+3.

Kategorisering Beskedent,Moderat, Højvækst bliver kun foretaget hvis virksomheden tilhører samme kategori i alle de seneste 3 år, dvs. at f.eks. højvækst eller beskedent i år1 og 2, men moderat i år3 resultere i NA. Kategorierne eksistere altså kun i deres "reneste form"

Forudsætning for at beregning foretages er at omsætning er steget min. 1 mio.kr. hvert i den seneste 3-årige periode}
}
\description{
Funktionen udregner vækstkategorien for henholdsvis omsætningsvækst.
}
\examples{
#data eksempel 

x <- 30000 # Omsætning i 2014
y <- 20000 # Omsætning i 2013
z <- 15000 # Omsætning i 2012
g <- 10000 # Omsætning i 2011

vaekst_oms(x,y,z,g)
}
\keyword{vækstkategori}

